Added card page
ура я нашёл как сделать коммит :D
