+++ src/components/CardDetails/CardDetails.css
@@ -1,0 +1,130 @@
+.window {
+    background-color: #282E44;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-weight: 700;
+    letter-spacing: 0.1em;
+}
+
+h1 {
+    color: #F8FBFF;
+    text-align: center;
+    font-size: 60px;
+    letter-spacing: 0.1em;
+    margin: 0.1em;
+}
+
+p {
+    margin: 0;
+}
+
+.description {
+    color:#F8FBFF;
+    font-size: 28px;
+    width: 80vw;
+    min-height: 35vh;
+    padding: 1.5em;
+    border: solid 4px #F8FBFF;
+    border-radius: 26px;
+    background-color: #212639;
+    margin: 1vmin;
+}
+
+.group-box {
+    display: flex;
+    flex-direction: row;
+}
+
+.p-group {
+    font-size: 18px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    margin: 3vmin 2vmin;
+    padding: 4vmin 6vmin;
+    border-radius: 26px;
+}
+
+.btn-group {
+    font-size: 18px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    margin: 3vmin 2vmin;
+    padding: 1vmin 6vmin;
+    border-radius: 26px;
+}
+
+.p-group {
+    background-color: #1B212F;
+    color: #F8FBFF;
+    width: 38vw;
+}
+
+.btn-group {
+    background-color: #F8FBFF;
+    color: #1B212F;
+    width: 28vw;
+}
+
+.p-group p {
+    padding: 0.6em 0;
+}
+
+.btn-group button {
+    font-size: 18px;
+    font-weight: 700;
+    color: #1B212F;
+    background-color: #F8FBFF;
+    border: none;
+    border-radius: 18px;
+    padding: 0.8em;
+}
+
+.btn-group button:hover {
+    background-color: #D2D2D2;
+}
+
+.btn {
+    font-size: 18px;
+    font-weight: 700;
+    color: #1B212F;
+    background-color: #F8FBFF;
+    border: none;
+    border-radius: 18px;
+    padding: 1em 4em;
+    margin: 2vmin 1vmin;
+}
+
+.btn:hover {
+    background-color: #D2D2D2;
+}
+
+hr {
+    width: 70%;
+    border: 1px solid;
+}
+
+.btn-group hr {
+    border-color: #D2D2D2; 
+}
+
+.p-group hr {
+    border-color: #464E65;
+}
+
+.btn-line {
+    display: flex;
+    justify-content: center;
+    align-content: center;
+}
+
+.close {
+    padding: 1em;
+    background-color: #F8FBFF;
+    border: none;
+    border-radius: 20px;
+    margin: 2vmin 1vmin;
+}
