+++ src/pages/Card.jsx
@@ -1,0 +1,12 @@
+import CardDetails from '../components/CardDetails/CardDetails.jsx'
+
+function Card() {
+
+    return (
+    <div>
+        <CardDetails />
+    </div>
+    );
+}
+    
+export {Card};
